fixup! Change the method the comparison function uses to check for inclusion
Change the comparision function so that rather than reducing the
array of hands which are beaten by a given hand to find if they contain
an element which is equal to the opponents hand, it now just uses the
includes method to check if the opponents hand is in the array.

Simple.