Replace function body of randomHand function
I found a more direct way of randomly selecting the keys in an
object. So, rather than generating a way of converting a number into
a string, to access the object by its key, I convert the hands object
into an array of the values for each key. I then randomly select the
value from the array.