Change randomHand to return an object, rather than a string
I discovered that my code doesn't work, and that's because it is
inconsistent in what type of object it uses to represent a hand.

The `compareHands` function expects two objects, containing a string
identifying which hand the object is, and an array of the hands which
this hand beats.

`randomHand` however, returns a string which is the key for the `hands`
object, which if you take the coresponding value from the `hands`
object, would give you the object which `compareHands` expects.

So, I've changed the result which `randomHands` gives to the object
form which `compareHands` expects.