Change the value `displayResult` is called with in the event listeners to the hand object form
I found that my code handles the representation for hands
inconsistently, and so didn't actually do what I wanted it to do.

I've changed the call to the `displayResult` function so that
instead of passing it a string of the name of the hand which the user
selects, we pass it the object in the `hands` object which is found
under the key named by the hand name.