Move the text for the hand selection buttons outside of the tags for the icons
The text of the hand selection buttons was too close to the icons
for my liking. The text was also inheriting the style of icons.

Moving the text outside of the icon tags seperates it from the icon
sufficiently and allows me to stlye the text without worrying about
inheritance from Font Awesome's library.