Add a meta tag declaring the viewport's settings
I've added the classic
`<meta name="viewport" content="width=device-width, initial-scale=1.0">`
tag to the head of `index.html`.

This is to make sure that the text and elements of the page scale
properly with the viewport's width. i.e. That text appears larger on
smaller devices and proporionaly smaller on devices with a larger screen
width.